Move TryGetString to MemoryMarshal, remove TryGetArray and update tests
