+++ src/System.IO.Pipelines/tests/TestMemoryPool.cs
@@ -96,7 +96,7 @@ namespace System.IO.Pipelines
             protected override bool TryGetArray(out ArraySegment<byte> arraySegment)
             {
                 _pool.CheckDisposed();
-                return _ownedMemory.Memory.TryGetArray(out arraySegment);
+                return MemoryMarshal.TryGetArray(_ownedMemory.Memory, out arraySegment);
             }
 
             public override bool IsDisposed
