+++ src/System.Net.WebSockets/src/System/Net/WebSockets/WebSocket.cs
@@ -35,7 +35,7 @@ namespace System.Net.WebSockets
 
         public virtual async ValueTask<ValueWebSocketReceiveResult> ReceiveAsync(Memory<byte> buffer, CancellationToken cancellationToken)
         {
-            if (buffer.TryGetArray(out ArraySegment<byte> arraySegment))
+            if (MemoryMarshal.TryGetArray(buffer, out ArraySegment<byte> arraySegment))
             {
                 WebSocketReceiveResult r = await ReceiveAsync(arraySegment, cancellationToken).ConfigureAwait(false);
                 return new ValueWebSocketReceiveResult(r.Count, r.MessageType, r.EndOfMessage);
